Refactoring from DependencyGraph to PCIGraph (just different naming for better understanding.
